// src/apis/models/FireBytes_Unified_PageFilters_Dtos_SavePageFilterAsPersonalInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FireBytes_Unified_PageFilters_Dtos_PageFilterDto } from './FireBytes_Unified_PageFilters_Dtos_PageFilterDto';
export type FireBytes_Unified_PageFilters_Dtos_SavePageFilterAsPersonalInput = {
    userId?: string | null;
    pageFilterDto?: FireBytes_Unified_PageFilters_Dtos_PageFilterDto;
};